requirements.txt
customtkinter==5.2.2
Pillow>=10.4.0
pillow-avif-plugin>=1.4.3
tkinterdnd2==0.3.0
